moving settings robots to their own pacakge
